This change addresses a side effect of a change done in Tomcat 3.2.3 to close a security hole.  In Tomcat 3.2.3 we disallowed any URI that contained encoded special characters like %, /, ., \.  While this did help close the security hole, it caused other problems.  First, it violates the URL specification (see RFC1630 and RFC1738) and second it prevents some useful URLs from working.
To address this I've modified the behavior of RequestUtil.URLDecode().  It
used to check for some encoded special characters and if it found them it
would throw and InvalidArgumentException.  I've added checks for the rest
of the special characters.  Also, instead of throwing an exception the
encoded character is left in the resulting string.

Thus a URI like /fu%2fba%72 will decode to /fu%2fbar.  Leaving the encoded
characters in the result will preclude them from matching against prefixes
or security constraints but the data will be available to servlets and JSP
pages that want the information, but they will have to decode these
strings on their own.

There is still an issue regarding the value returned by getPathInfo().
The specification states (in the 4/27/2000 errata) that getPathInfo()
should return a decoded value.  The value currently returned by
getPathInfo() does not decode any of the special characters listed above.
It is not clear to me now if we should take the extra step of completely
decoding the path info or not, so for now I'm leaving it alone.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@284250 13f79535-47bb-0310-9956-ffa450edef68