Porting change from tomcat_32.
Fixes Bugzilla 2333.  Instead of the proposed patch, which did its own
memory allocation, I used jk_pool_strdup(), but the end result is the same.

In the process of fixing this one I noticed that the jk_pool created in
mod_jk was not being closed.  This usually doesn't leak because the pool's
buffer is allocated on the stack in jk_handler(), but it would leak any
dynamically allocated memory.

PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/tomcat@295828 13f79535-47bb-0310-9956-ffa450edef68