Impove error handling when the user has choosen not to define a ROOT context.
A better solution is to force the definition of DEFAULT@ROOT, but that creates too many problems this close to release.  I'm the first to vote to role this back once someInterceptor takes responsibilty for DEFAULT@ROOT.

It is very easy (and, it seems common) to setup Tomcat without a ROOT context. This caused Tomcat to generate useless exception reports in the log.  With this patch, you get the expected 404 response to an un-mappable request.

This fixes Bug #3988
Reported by: Alex Polverini polverini@explo-it.com
Reported by: Hans Schmid hans.schmid@einsurance.de
PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/tomcat@284257 13f79535-47bb-0310-9956-ffa450edef68