Replace SimpleHashtable ( the unsynchronized hashtable ) with a real hashtable.
The prefix and exact mapping can change at runtime ( we dynamically add mappings
for jsps and /servlet ). We could avoid the sync, but it's safer to use Hashtable for now.

Regarding *hosts, the keys() can't be used in multiple threads - so we have to use
SimpleHashtable ( or possibly move it to an [] and avoid creating Enumeration and
using a hashtable when in fact we check all elements ).

Most of the time this isn't a problem, but under hight load you may see exceptions.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@284290 13f79535-47bb-0310-9956-ffa450edef68