Add a simple non-Javascript-based tree control, implemented as a custom tag (implemented here because I could not find a good general purpose one that met the requirements -- it's not tied to Tomcat so will likely be abstracted out later, but is here for development convenience for now.)
A test page allows you to exercise the tree (either to expand/contract
nodes or to select a particular element):

  http://localhost:8080/admin/tree-control-test.jsp

(You will need to set up a user with the "admin" role in order to log on
to the admin app to exercise this page).

Note that the visual appearance of the tree nodes can be influenced by the
choice of images used, and by the CSS style classes that you select.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@286988 13f79535-47bb-0310-9956-ffa450edef68