Update to the new API.
Note that the .o is generated using the same rules as javac - in a separate directory
( the buildDir attribute ), using a package structure derived from the file location.
( i.e. the file location minus base dir in <src> ). This avoids potential
name conflicts, when same name is used for .c files in different directories.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@295951 13f79535-47bb-0310-9956-ffa450edef68