Set the URLStreamHandlerFactory to allow webapps to create https URLS.
With this servlets can create a java.net.URL("https:/...") with JSSE/OpenTLS either installed not on the system ClassLoader, or missing altogether.

I also added a (very dangerous) shutdown method, to aid developers who want the "advanced" features.  It is never called from a normal Tomcat.  It is only for people that embed Tomcat in other applications.  I'm agreeing with Remy here that adding an API that should have been there all along is not that big of a deal.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@284366 13f79535-47bb-0310-9956-ffa450edef68