Removing a lot of duplicated code, update everything.
I tested with both 4.0.1 and 3.3 - it seems to work fine with both 13 and 14,
but I probably left few debug statements in.

This is not 'complete', there are still many things to clean up and make
sure we can use mod_jk ajp14 with both 13 and 14 servers and to start
integrating the 1.4 callbacks into 4.0.1.

Based on the feedback on the 'jk_handler' proposal, the code will be greatly
simplified and enhanced in the next few weeks.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@295994 13f79535-47bb-0310-9956-ffa450edef68