Merge logic from Ajp14 in Ajp13, removing the duplicated code.
The difference is that if Ajp13 is configured with a 'setBackward( false )' it'll
require an auth phase and support all additional handlers ( negotiation, etc ). Otherwise it'll
behave as it did before, as a 'basic' ajp13 worker ( important for 3.2.x where 14 is unlikely
to get ported, and 4.0 where 14 is not yet completed ).


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296001 13f79535-47bb-0310-9956-ffa450edef68