Added new properties for apr.
Make sure apr-related files are excluded if apr is not available. This will
be enhanced as we go.

Note that short-term we'll keep the non-APR files. Some features ( like use of
shared memory or pipes for IPC, etc ) will not be available, but the basic
AJP over TCP/IP will work in any case.

As we start using APR we'll try to eliminate as much 'unportable' code as possible.
When this is done, we can get rid of configure and use only a simpler build system
( configure is great for detecting system-specific details, but we'll have less
need for that since APR provides the needed abstractions ).


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@296022 13f79535-47bb-0310-9956-ffa450edef68