Fix Spec compliance problem with jsp_precompile.
This was reported against 4.x, but we're just as broken.  There is the additional problem that the 1.1 spec actually specifies two mutually exclusive behaviors to follow.  Since this section of the 1.1 spec and the 1.2 spec are mostly the same, I'm following the 1.2 spec as a 'clarification' of the 1.1 spec.  This means that "jsp_precompile=false" still is not passed to the jsp page.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@284381 13f79535-47bb-0310-9956-ffa450edef68