More fixes, now it seems the startup sequence is ok.
The support for urimappings.properties ( used right now only for IIS )
is now available to all servers, and can be specified in workers.properties as well.

( next step is to have an easy way to specify and load multiple files ).

Note that worker.properties settings can be replaced with JkWorker name value in
httpd.conf ( i.e. you can do all the settings in httpd.conf, no need for a real
worker.properties if you prefer to use a single file )


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296049 13f79535-47bb-0310-9956-ffa450edef68