Consolidate data structures.
A lot of 'ajp private' information is very usefull for other workers as well.
Having it exposed in the main structure will make things simpler, and eliminates
a lot of duplication.

We do lose so 'data hiding' and private fields, but I think it's worth it.

( for example we want the jni worker to be able to use the same
discovery as the ajp14 worker )


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296060 13f79535-47bb-0310-9956-ffa450edef68