The discovery handler.
There are few issues with the autoconf - like the fact that tomcat must be
started, the handling of load balancing, etc.

The whole configuration must be re-opened, we need a solution that scales for
complex sites. This is just a start.

The good news is that this is optional, so we'll worry about discovery after
 we get everything else fixed.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296069 13f79535-47bb-0310-9956-ffa450edef68