Added the initial stub for the jni code to use apr from java.
This is needed to implement any 'fun' channel ( unix domain, shmem, etc ). It'll
also be used to implement signals ( so Jk can triger soft restarts on apache )
and maybe change uid ( not related with jk, but trivial to do ).

For now the focus is getting the functionality in ( i.e. the JNI code ), not
the interfaces - long term we should use some nice interfaces ( like
Socket for unix domain sockets, etc )


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@296098 13f79535-47bb-0310-9956-ffa450edef68