One more thing extracted from Henri's discovery, plus things merged from jk_uriEnv ( merged from various places like apache private structs ).
jk_webapp coresponds to Context, it should have all per/context settings.

Note that now the granularity for setting the worker is bigger - you can
set the jk worker per context only. ( we can't make it per uri, but I'm
not sure it makes sense, and it's easier this way ).


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296101 13f79535-47bb-0310-9956-ffa450edef68