- use r->request_config to pass the uriEnv ( the result of mapping ) instead of a note. This way we avoid few ( dozens ? ) string cmps, no need to pass the workerEnv (we pass the real structure ), the code is much simpler.
- Finally ( ??? ) get rid of the double initialization. Apache unloads all
modules after it reads the config, so if .so is used our 'is_initialized'
will disapear as well.

Now we use the 'top' pool ( gpool ), which is unlikely to disapear, and
we set a userdata attribute. We do the initialization only the second time.
( we should validate the data on the first run ).

( this still needs testing, but it seems it is a good track )


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@296113 13f79535-47bb-0310-9956-ffa450edef68