Finish jk_objCache, start using it in jk_worker.
( the code is moved from ajp13_worker ). I added it to lb_worker, not yet to
jni - but it may be usefull in few other places as well.

Make sure the 'per connection' pool in endpoint is different from the
'main' pool ( that has the same life as the endpoint ).

It seems the objCache code is a bit inefficient, I don't know why it works
the way it does ( it could just be a stack ).


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296115 13f79535-47bb-0310-9956-ffa450edef68