Better abstraction for message serializing/deserializing, protecting the internal data ( the buf is no longer used/useable outside ).
This is supposed to allow other encodings/protocols to be added with minimal
effort and without changes in mod_jk.
In addition, this is supposed to work with JNI as well - each 'appendFoo' can
be translated to converting to java and saving it in a jobject[], with
send() implemented as a jni call.

It also cleans up a lot of code that used 'dirty' tricks.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296121 13f79535-47bb-0310-9956-ffa450edef68