Removed most of the stuff from ajp14, moved to individual handlers.
We must start making the distinction between 'protocol' and 'api'.

jk supports 'ajp13' protocol, 'jni' protocol, etc. It could also support warp
protocol or even a (minimal) IIOP.

The API is the set of messages we use - confusingly called 'ajp13', 'ajp14', etc.
All the details about message is implemented in message handlers, which are
independent ( or should be ) of the protocol.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296126 13f79535-47bb-0310-9956-ffa450edef68