JkMounts will be added to the config, but no actual call is made before init().
( the same model that is used on the java side in both 3.x and 4.0 - first we
set properties, then we properly start the object )
No distinction is made between JkMounts or uriworkers.properties or even
properties in workers.properties ( or any other properties file ).

During the config stage, we'll collect properties from various sources:
- workers.properties.
- native server config ( both specialized JkMount or specific options, or
JkWorker foo bar which is equivalent with foo=bar in workers.properties ).
- urimap.properties

That means the config file to be used is entirely matter of taste.

Few more fixes to the code that detects the 'first config' done by
apache and doesn't initialize on it.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296139 13f79535-47bb-0310-9956-ffa450edef68