Add unit tests for RequestDispatcher.include() for all combinations of: * Servlet or JSP includer * Servlet or JSP includee * Flush or no-flush before include
Currently, the RequestDispatcher implementation has a bug (no output in
the response for the include or anything after it returns) whenever an
include is performed after a flush (either because an including servlet
called response.flushBuffer(), or because an including JSP page is
exectuing a <jsp:include ... flush="true"/>).  This manifests itself as
failures on the corresponding tests in the "Jsp" and "RequestDispatcher"
targets.

I suspect this bug was introduced during the recent refactorings for error
page handling -- alas, our unit tests did not catch it :-(  I've looked
for a while and can't find it yet.  Any ideas?


git-svn-id: https://svn.apache.org/repos/asf/tomcat@287210 13f79535-47bb-0310-9956-ffa450edef68