The map implementation using apr_table.
One problem is that apr_table supports only string values. We use
string values for most of the stuff, but we need void * for uriMap.

But in uriMap we should better use hashtable - and maybe we shuld have
2 interfaces ?

For now aprtable will be used only for headers/env where we know
we only have strings, and the 'original' jk_map will continue to
be used.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296155 13f79535-47bb-0310-9956-ffa450edef68