actually send post data beyond the first ~8K to the servlet container.
if a client request contains post data, the first chunk of it is sent to the
container directly after the forward request is sent.  if not all of the data
fits in the first chunk, the container will get subsequent chunks by
sending the connector a GET_BODY_CHUNK message.  this checkin fixes the
reponse to the GET_BODY_CHUNK message so that the connector will sent the
container more post data.

i believe this also fixes bug 4223.
PR: 4223
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296175 13f79535-47bb-0310-9956-ffa450edef68