Don't check dates if the static page is part of a jsp:include
This was reported as Bug 5497 against 4.x.  In the report, it only happened because of a buggy browser, but there are still possible (if rare) for a servlet to set the Last-Modified header.  In any case, we can't send a 304 response even if we wanted to for an included page.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@284417 13f79535-47bb-0310-9956-ffa450edef68