The statusHandler is returning a body for 304 responses, which is forbidden by both HTTP/1.1 and HTTP/1.0.
IE is forgiving, but Netscape Navs < 6.x assume that a
body is not there, even if the content-length is set
(it is).  So, the 304 response body gets munged with the
next response body.  Of course, this does not happen
standalone because the bug requires connections to be
kept alive.

cf <http://www.w3.org/Protocols/HTTP/1.0/spec.html#Code304>,
   <http://www.w3.org/Protocols/HTTP/1.1/spec.html#Code304>


git-svn-id: https://svn.apache.org/repos/asf/tomcat@284419 13f79535-47bb-0310-9956-ffa450edef68