JVM Temporary Directory (java.io.tmpdir property) security patch ----------------------------------------------------------------
The java.io.File.createTempFile() methods use the property
java.io.tmpdir for the location of the directory where the JVM
places temporary files.  If this property doesn't exist,
/tmp or /var/tmp gets used on unix systems or c:\\temp on WIN32 systems.

This default directory is the same directory that all applications
can use for temporary files.  Sharing the same system temporary
directory between Tomcat and other applications could be a security
risk if sensitive data is stored there by other applications.

If you run Tomcat with the SecurityManager enabled you would have to
grant Tomcat a FilePermission with read, write, and delete for the
temporary directory.

This series of patches enables the use of a temporary directory
located at $CATALINA_BASE/temp so that Tomcat's use of the temporary
directory is isolated from the rest of the system.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@287260 13f79535-47bb-0310-9956-ffa450edef68