"New" jk. The code is cut&pasted from the old jk, with a bit of cleanup and reorganization to match the object model on the C side.
The most significant change is the removal of String from MsgAjp ( which
is the old AjpPacket ). Jk doesn't have enough knowledge to deal with
strings - the higher layer ( the connector ) should do the byte-char
conversion, we just marshal byte[], int.

The main reason is that charset encoding is hard to detect at the packet
level ( it can be in the Session, or extracted from headers, or
set by user in 2.3 ). The old code was not working well with different
charsets.

The second reason is performance - doing String.getBytes() proved to be
very inefficient. The connector should use C2B or B2C or JDK1.4
to do the conversion.

You may notice ChannelSocket, which now does send/receive the messages
( part of the old Ajp13 object ). The intention is to add other channels,
using different transports ( unix sockets, etc ).

Msg abstracts the encoding, and again we may have a different impl.
in future ( but less likely, the current ajp13 encoding seems fine,
maybe we can experiment with alignment or CDR ).

Endpoint represents the point to point connection between java and c sides,
it's very 'light' on the java side ( only generic notes ). In future we
should simplify the C endpoint too.

Worker is the link to the higher layer ( tomcatXX ). It is no longer the
factory for endpoint, etc - the endpoint/channel/msg are now a bit
more independent ( i.e. in java ), to avoid duplicating code.

Of course, nothing is final :-) It kind of works now, but needs more testing,
and the objects are certain to change when we'll have the first
'alternative' transport.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296195 13f79535-47bb-0310-9956-ffa450edef68