This is the catalina-specific servlet.
There are many ways to plug the connector. Using the servlet allows the config
and deployment to use the typical 'webapp' pattern. JkServlet will detect
the container and use a container-specific code to plug itself in.

In the case of 4.0, the only way for a priviledged app to access the internals
is via a servlet implementing an interface and belonging to a o.a.catalina package.

It would also be possbile to add a different ( and cleaner ) mechanism
by using a valve - but at least for the released 4.0 that's the simplest
way ( at least for the user, who'll have minimal setup to do ).


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296218 13f79535-47bb-0310-9956-ffa450edef68