This is an extension of C2BConverter, with support for converting strings and message bytes.
One of the biggest problems in jk1 was the fact that all communication was done
using String.getBytes() and new String(bytes[]). This is extremely inefficient.

Now Msg has only support for adding Bytes to the message - the higher layer is
responsible with the conversion. Msg has no way to guess the encoding anyway -
it could be in the Session or set by some interceptor/valve.

I hate doing this, but for now I think it's better to have no method to add
a string to the message - this will force all code to deal with the encoding.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296222 13f79535-47bb-0310-9956-ffa450edef68