Added code for unix sockets ( cut & paste from the C version ). It'll be replaced/merged with jfc's version.
I also added few other methods to AprImpl, now pools seem to work ( it seems the first
step in using APR ). Unix sockets are using 'native' code, not APR, but I hope
it'll be integrated with apr code and we'll replace this with the plain apr socket.

Other usefull functions we should add and test from APR ( for the short term ):
- send signals
- change user id

Probably shared memory will happen in jk3.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@296229 13f79535-47bb-0310-9956-ffa450edef68