a quick patch to attempt to shorten shutdown time, plus some additional logging.
the idea is simple:  set a flag when the processor is handling
a request, then when stopping, only wait on the worker
thread if a request is currently being handled.

this patch won't make a difference if the connector is busy and
there are a bunch of processors currently handling requests, but
it will make a difference if the connector is not busy.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296242 13f79535-47bb-0310-9956-ffa450edef68