This is the code extracted from jni worker.
Few changes - we do now a lot of 'guessing' to avoid asking the user
for things we can find for ourself. The defaults do work now fine - it
required LD_LIBRARY_PATH to be set before starting apache ( that's required
by jni ), and if you also set TOMCAT_HOME and JAVA_HOME ( that should
be already set or easy to do ) jni will figure everything it needs.

The old settings in worker.properties can be used to override the guessed
values, but the nightmare of setting jni worker should be close to end.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296264 13f79535-47bb-0310-9956-ffa450edef68