Updated to write a "properties" file in addition to registry file.  This file will be written to the same directory as the registry file.
Add an isapiRedirector attribute.  This attribute specifies the name of
the redirector DLL. It affects the value of the extension_uri configuration
setting and the name of the "properties" file written.  This will make it
easier to use the isapi redirector in jakarta-tomcat-connectors which
has a slightly different name as the one in jakarta-tomcat.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@284470 13f79535-47bb-0310-9956-ffa450edef68