Change to use JarFile rather than JarURLConnection.
Continuing to work around the problems with Java 1.3.x locking up jar files.  Now, the jar file is closed after defining the package (and so can be replaced w/o waiting for finalize).  Now, it should be possible to have both the good reloading of 3.3 with the brain-challenged loading behavior of 3.2.  This is about as non-critical as code in 3.3 gets, so the extra open/close shouldn't hurt perfomance at all.

Also some formatting fixes.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@284476 13f79535-47bb-0310-9956-ffa450edef68