Update MBean creation support so that MBeans are created in all cases where new components are added dynamically, by registering appropriate event listeners.
TODO:  Creation of MBean object names for valves is not returning unique
names, so only the last valve for a particular component is done.

TODO:  Looks like we need to add events for StandardServer.addService() and
StandardServer.removeService() so that ServerLifecycleListener can listen in.

TODO:  Need to flesh out removing listeners and MBeans as components are
removed (should be pretty much a mirror image of the times when things are
created or registered).


git-svn-id: https://svn.apache.org/repos/asf/tomcat@287389 13f79535-47bb-0310-9956-ffa450edef68