A number of bug fixes - make sure we save the global ref, not the ref. We now get an Endpoint from the java side - and cache/reuse it with the jni endpoint.
The jni channel is 'interesting' - it's the first non-stream channel. It could
be treated as a stream, by using 2 threads ( so send/receive model will work ),
but I wanted to preserve 'single thread, no sync' model from the previous
jni worker.

It may seem a bit complicated - and it adds some limitations on the model, but
I think it's worth it.

The idea is that, as before, the first message ( containing the request ) gives
control to tomcat who may send back messages ( and get back responses to it's
messages ). The send() method in channel is doing exactly this first step.

Tomcat will use a native method to send messages ( that replaces receive(),
which is not used ), which are dispatched. The response is actually put in
the same buffer - a single jarray pin is needed.

We must make sure we don't run into buffer problems - but that can be resolved.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@296318 13f79535-47bb-0310-9956-ffa450edef68