That's one of the biggest changes.
First, the endpoint management is now more explicit and direct - since this
worker is sending the request, it needs endpoint and it recycles it.

Second, we treat jni as a particular case - with a different way to handle it.
What's nice is that now almost all of the code is shared and common - we do get
reused endpoints for jni and all the good stuff from ajp ( including efficient
c2b and less GC ).


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296323 13f79535-47bb-0310-9956-ffa450edef68