Removed ChannelUnixSocket ( with JFC's agreement ). It's duplicated in ChannelUnSocket.
Removed HandlerEcho, it was empty anyway. The first step is to get the original
ajp13 working, in a future release we can add other handlers and messages
( sorry Henri, most of the new stuff in ajp14 is not yet added, I'll go back to
it after I finish ajp13 and jni stuff )

JkChannel - I moved back to Channel as abstract class, it easier for now.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296335 13f79535-47bb-0310-9956-ffa450edef68