Make it possible to build the HEAD branch of Tomcat 4 against the newest Xerces 2.0.0beta4 distribution (which uses two files, "xmlParserAPIs.jar" and "xercesImpl.jar"), or against an older distribution (which uses one file, "xerces.jar").  The choice of which implementation to use is defined by setting one or the other of the following sets of properties:
* For Xerces 1.3.1 up through and including 2.0.0beta3,
  set "xerces.jar" to point to the full pathname of that file.

* For Xerces 2.0.0beta4 or later, set "xmlParserAPIs.jar" and
  "xercesImpl.jar" to point to the full pathnames of the
  corresponding files.

If you have both sets of properties defined (which will be common if you
are building lots of packages that depend on a "xerces.jar" property), the
2.0.0beta4 (or later) version of Xerces will be the one used for the Tomcat
build, and will also be the parser that is copied in to the common/lib
directory.

At the moment, I would lean against porting this mechanism to the 4.0.2
release, because it appears we will release 4.0.2 before Xerces 2.0 is
finalized.  IMHO, Tomcat 4.0.2 should be built against, and include, the
stable Xerces 1.4.4 release.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@287425 13f79535-47bb-0310-9956-ffa450edef68