Reworking the "delay" kluge for test software in which client and server run on the same system and experience "Socket reset by peer" errors due to unfavorable sequencing of thread execution.  This update was needed because adding the sleep(0) caused the failure rate of Tomcat 3.3 tests under Cactus to go from infrequently to frequently on the primary Gump system running Linux.
Changed delaySocketClose to closeSocketDelay and made the attribute
numeric.  The default is -1, which adds no delay.  A value of 0, adds a
Thread.yield() prior to closing the socket.  A value greater than 1 causes
CPU cycles to be eaten for the specified number of milliseconds prior to
closing the socket.  Using Thread.sleep(n) proved to be unhelpful on
RH Linux 7.1.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@284505 13f79535-47bb-0310-9956-ffa450edef68