Make certain that the Writer isn't in the error state.
This is the quickest and safest fix.  We could try for something better in 3.3.2 if we want.  However, this almost never happens (since usually an IOException kills the entire thread), so there really isn't that much extra garbage to collect.

Fix for bug #6234.
Reported by: Pete Fischer pfischer@bbn.com


git-svn-id: https://svn.apache.org/repos/asf/tomcat@284523 13f79535-47bb-0310-9956-ffa450edef68