Renamed Handler to JkHandler - the name is hugely overloaded, and likely to create conflicts.
I think the architecture of jk2 is close to beeing finalized, the current
model should be powerfull enough for most of our needs. I copied
a lot from axis, which provides an extremely powerfull and flexible
model for message processing ( which is what jk2 is doing as well ).

JkHandler is the 'core' of the whole model - it supports the classic
'hook' pattern used in most apache products. Given that our
preferences are split, JkHandler supports both catalina-like
'recursive' hooks invocation ( using next ) and apache ( or 3.3 ) style
'iterative' invocation.

I've also renamed endpoint to MsgContext. The signature is different
from Axis's handler, in that we pass both the 'source' message and the
MsgContext. The reason is that Jk should be able to handle 'internal'
messages during the processing of the main message. The thread local
current MC is hacky.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296360 13f79535-47bb-0310-9956-ffa450edef68