Dispatch handler - based on the message type will call the right handler.
This is specific to ajp13 ( or any other transport that stores the message type
as the first byte of the message :-).
Each protocol is supposed to have a set of protocol-specific handlers,
some may be generic.

What we'll eventually do is create an HandlerAjp13 to consolidate ajp13
specific processing. Other protocols could do the same.
Then whatever info containes the 'target' will be stored as the message
tag, and the rest of the info will be decoded in message/request/etc.

Note that the tag is an int - we want to reduce the use of Strings,
and do table dispatching, so protocol handlers should use a message handle


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296367 13f79535-47bb-0310-9956-ffa450edef68