Update for the simplified model ( handler-centric )
Read conf/jk2.properties and use if for configuration. It is very
similar ( or identical ) with workers.properties - we want the C code
to read it easily ( same code will be used ), that simplifies a lot in
the model and eliminates duplication of config ( right now you need
to change both server.xml and workers.properties ).

If we want to use one of the alternate config mechanisms, we'll
generate jk2.properties to get the same effect.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296374 13f79535-47bb-0310-9956-ffa450edef68