Moved loadJsp from LoadOnStartupInterceptor to ServletHandler.
There are too many ways to access a declared JSP, so centralize the hook to make certain that it gets compiled.

In 3.3.2 we should probably add a 'prePreServletInit' hook to the API so that JspInterceptor can take care of these cases on it's own.

Fix for bug #6348
Reported by: Vincent Massol vmassol@octo.com


git-svn-id: https://svn.apache.org/repos/asf/tomcat@284543 13f79535-47bb-0310-9956-ffa450edef68