Enhance the Manager webapp to support two new facilities:
* Deploy a new application (optionally to a Tomcat instance running
  on a different server) by uploading the WAR file to it.

* Undeploy an existing application and remove its document base directory
  from webapps.

* Support these commands with new custom Ant tasks, typically named
  <deploy> and <undeploy>.

The first of these commands is useful only within tools, becasue it requires
an HTTP put.  However, it is supported by an Ant task for use in build
scripts.

If you want to do application development "in place", inside the context
root directory of a running webapp, you will probably want to use the
"install" command rather than the "deploy" command.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@287558 13f79535-47bb-0310-9956-ffa450edef68