Modify the way that the Common class loader is created to include the following contents:   Unpacked classes in $CATALINA_HOME/common/classes   All JAR files in $CATALINA_HOME/common/endorsed   All JAR files in $CATALINA_HOME/common/lib and move the Xerces parser we install from "common/lib" to "common/endorsed".
Background:  JDK 1.4, among many other changes, packages a version of Xerces
and makes it accessible via the JAXP APIs.  However, because these are now
classes in the Java runtime, you cannot normally override this with your
own parser.  However, a mechanism is provided (setting the system property
"java.endorsed.dirs" to point at a particular directory -- which we'll set
to "$CATALINA_HOME/common/endorsed" in the startup scripts -- that replaces
the embedded parser with the new one (such as Xerces 2.0).

TODO:  Update the shell scripts to set the endorsed directory.

TODO:  Update the class-loader-howto doc to describe all of this.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@287572 13f79535-47bb-0310-9956-ffa450edef68