Finally, the config changed we discussed.
There is only one config directive for mod_jk that we should care:
  JkSet name value

JkSet workersFile  ... will read properties from a file, you can
use it multiple times.

Also "JkSet workersFile urimap.properties" will read a file in the
style used by IIS ( /uri=worker ).
Or
  JkSet /examples ajp13
will do the same as the old "JkMount /examples ajp13" ( or "/examples=ajp13"
in any properties file ).


Properties file will hopefully be automatically generated, and this allows
httpd.conf to be used in the exact way with workers.properties ( and reverse).

The other special directive is JkSetUri, in a <Location> context, which will
set properties on a uriEnv.

I'll add an equivalent thing, properties-style, and that should be it.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@296452 13f79535-47bb-0310-9956-ffa450edef68