Restore the unpackWARs behavior (but only for WAR files in the appBase subdirectory) that was broken by previous changes to deployment and the Manager webapp's servlet.  The WAR-expansion logic was refactored from StandardHostDeployer to HostConfig, which is the only place that requires it.
Applications that are added via the Manager webapp's "/install" or "/deploy"
commands are *not* auto-expanded -- if you want your webapp to run from an
unpacked directory, you should install an unpacked directory.

NOTE:  You can also get auto-unpack service by simply dropping a WAR file
into the appBase directory while Tomcat is running - a background thread
will notice this and unpack it in the same way that startup works.

NOTE:  As before, any auto-unpacked directory is *never* deleted, because
the user has probably changed things and would be very annoyed if it was
removed.

I'm going to make a pass through the JavaDocs and the documentation files
to make sure that they are all up to date with respect to the actual
functionality.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@287706 13f79535-47bb-0310-9956-ffa450edef68