Finally remove validate().
Initialization is done using setProperty() followed by an init().
The worker should be independent of the way it is configured -
and 'smart' workers can change some settings at runtime.

Adding individual setters and introspection is of course not
easy in C, but setProperty() is close enough for what we
need.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296457 13f79535-47bb-0310-9956-ffa450edef68