Minor fix to the below changes.
This fixes a spec compliance bug, SRV.3.7.1 states that the
tempdir for a Servlet Context must be unique.

Added the Engine name as a directory path component when creating
the default work directory for a web application.  This fixes a bug
where the same work directory would be used for the http and https version
of a virtual host.

Added the workDir attribute to the Host configuration.  This makes it
easier to set the base directory where the work directories for
applications are placed when virtual hosting sites with Tomcat.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@287719 13f79535-47bb-0310-9956-ffa450edef68