Merged changes from 3.3
Addition of stop_cmd, allowing the stop to be done by executing
a command. ( catalina shutdown protocol is already supported )

Removed some of the checks for settings that are not required.
If cmd_line is specified, there is no need for server.xml
or other tomcat-specific things.

cmd_line, stdout, stderr are required.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296527 13f79535-47bb-0310-9956-ffa450edef68