Refactor the TLD scanning code in ContextConfig to improve clarity, avoid redundant parsing of the same TLD, and some wasted effort trying to parse a JAR as an XML document (and vice versa) in the old code.
Currently, the jarFile.close() statement to close a JAR file after scanning
is still commented out (as it was before).  Closing the JAR file here seems
to mess up the WebappClassLoader (which has already included this JAR file
in its repository of available URLs) by causing "Zip File Is Closed"
exceptions when it attempts to load classes from that JAR.

In HostConfig, the only change was to close the JAR in a finally block when
expanding a WAR, even if an exception has occurred.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@287770 13f79535-47bb-0310-9956-ffa450edef68