Modified to match the new config mechanism.
The configuration is set using the jk_bean, which is managed by jk_config.
In theory jk_config doesn't need to know about the details of the object -
the name of the module and the name of the property is enough.

I also made the init() more consistent ( both internally and with the java
patterns ).


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296555 13f79535-47bb-0310-9956-ffa450edef68