Update workerEnv for the new config system.
Note that object creation and management is now centralized in jk_env/jk_config.

WorkerEnv, uriMap are no longer creating workers, uris - instead
of createWorker() we now use addWorker().

This is simpler ( == less code ), but require a bit more attention to
the initialization ( i.e. we may have to add a 'state' information, and
pay attention to the init ).

Another problem is that when a worker is created in a generic way, it
is harder to register it with the workerEnv. One solution is to
have worker-specific code that will call addWorker, another is to
have the worker register itself ( by calling getInstance( "workerEnv ") )


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@296562 13f79535-47bb-0310-9956-ffa450edef68