Update lb worker.
Changing the list of workers at runtime is now possible ( or
close enough :-)

The lb worker is no longer creating the workers itself, the
config/env modules are responsible with creating/config of
jk components.

The lbfactor is no longer 'pulled' by lb_worker, but is
set on each worker that may participate in a pool. ( ie.
forwarding workers - jni or ajp ).


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296567 13f79535-47bb-0310-9956-ffa450edef68