Don't stop the processing if the response is already committed.
Letting the IllegalStateException through was preventing error handling from working if e.g. the servlet throws an exception after the response is committed.  This in turn prevents the useful stack trace from showing up on development servers.  Under normal use, the response should never be committed at this point on the primary request.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@296578 13f79535-47bb-0310-9956-ffa450edef68