Clean up, reorganization, implement missing features.
The code calling env must get an env ( at least once per thread ). The env
object is recycled and includes a tmp pool ( to be used for temp allocations,
the pool is recycled when the env is released ). It also includes minimal
support for exceptions, JNI style. ( not used right now )

The code to create and manage instances is more consistent and hopefully
cleaner.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@296621 13f79535-47bb-0310-9956-ffa450edef68