The handler is a generic mechanism to be used to implement both recursive and iterative chains ( i.e. Interceptor-like or Valve-like ).
It will be used for Jk and hopefully Coyote to keep things
modular.

Please review and make the apropriate changes - but try to keep it minimal
and independent of any protocol/special use.

The model I had in mind was the Hook used in Apache2/APR, as well as
Axis Handler ( besides Interceptor/Valve/JkHandler )


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296651 13f79535-47bb-0310-9956-ffa450edef68