Improve the Thread-safety on the 3.3 Session exipiration.
I was seeing every so often that a brand-new session was getting expired.  With this we take a snap-shot of the sessions so that we don't have to worry about the data changing under our feet while we're doing the checking.  By keeping the array as part of the Object, we should avoid excess GC as well.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296658 13f79535-47bb-0310-9956-ffa450edef68