Initial implementation of http11 'protocolHandler'
The SSL stuff is not completely plugged in, and it will require a recompilation
of util ( so you have the util.net available ).

It also seem to have some problems with the output ( but more likley the bug
is in my refactored connector ). The code will be cleaned up after all pieces
are assembled.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@296667 13f79535-47bb-0310-9956-ffa450edef68