If host header is not set, use the local socket to set the local port and hostname.
Remy - I'm not sure if this is correct, but I see no other way to
pass this info ( I could add a callback - but this is needed in most
calls ).

If there is any particular reason to keep port and localhost unset
I'll rollback and add the callback.

( this used to be in the 3.3-specific code, and is needed to construct
redirects, etc )


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296713 13f79535-47bb-0310-9956-ffa450edef68