A small Log impl. on top of common-log, to get all logs go to the same logger. It's not set by default.
Various components use Log and some are using commonns-logging. It seems
commons-logging is used in the new stuff and I think it should be recommended
for all new code.

However 3.x and some utils are using this Log interface. This class will
allow old code to have the log go to the new log interface.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296752 13f79535-47bb-0310-9956-ffa450edef68