A bigger change in the way we build native2.
So far we had a jni_connect ( I changed the name to libjkjni since it conflicts with jk1 )
It contained a subset of jk and the native methods ( the minimum amount required to get
jni working ). I changed this to include the whole thing - that will allow java to
access all the features of jk, including the future scoreboard and all channels/etc.
( and more APR features will be used/exposed directly or in jk objects at a higher level)

A second change is in the mod_jk2.so - it didn't include any jni method. That's
a source of big problems since the jni worker will try to load jni_connect and is
likely to run into conflicts.

The JNI methods do not have _any_ dependency on java - just plain C methods taking
some pointer parameters. The JNI header and the pointers to functions are doing
all the magic. So if we include support for the JNI worker ( i.e. we have jni.h )
we can also include the jni methods in the same .so


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@296762 13f79535-47bb-0310-9956-ffa450edef68