Got unix socket to work again.
Note that we still use 'direct' unix calls, not the APR wrappers ( is it
finished ? I read about adding it in the plain apr sockets, but didn't see the
code yet ).

I also changed the names and messages to show the real source ( the code was
cut&pasted from channel_socket ).

Support the same 'trick' of using the object name as the constructor/URL.

All you need is:

[worker.ajp13:apr]
channel=channel.apr:${jkHome}/work/jk2.socket


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296763 13f79535-47bb-0310-9956-ffa450edef68