Got it to use plain main() and a much simpler and intuitive syntax ( I hope ). At least it's using standard java options instead of invented jk option names
Both Tomcat3.3 and 4.0 are starting fine - and seem to find all the native
libs they need ( if LD_LIBRARY_PATH is set ). There is one more problem
with setting up the ChannelJni ( and detecting and failing back gracefully ).

Then the big problem will be getting shm to work and dealing with the
real difficult reconfiguration of workers when a new apache process
is started ( sort of a dynamic pool of workers - with apache monitoring
and managing creation of workers ). That may not happen completely
for this release ( except maybe on windows or if a single process is used ).


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@296811 13f79535-47bb-0310-9956-ffa450edef68