Few changes to MsgContext to allow it to be used in a more flexible way.
Most of the methods can become generic 'notes' ( but we need
to support 'long' notes, I wouldn't like to convert to Long ).

MsgContext now has a 'type' ( from o.a.t.u.handler ) - that'll
eventually replace the Action and be commont to all coyote components.

We also store the 'source' - the handler that originally
initiated the callback ( to be used to send messages to the other
side ) and 'next' - the next handler in the chain.

All the changes are cleanups for jni and getting the code
closer to the target.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@296844 13f79535-47bb-0310-9956-ffa450edef68