Added shm native calls.
Cleanup the interfaces - now use the normal JkHandler interface for
configuration.

AprImpl must be loaded in the main classloader, with common-logging and
jk.core in order to get JNI-channel to work ( that will be fixed later, it is
also possible to get around the class loaders by pushing the objects via
native calles instead of loading them with FindClass )
( normal jni will work anyway ).

For callbacks we use normal JkHandler and MsgContext - it
coresponds to an apache hook and behaves in a similar way. ( ok, the
concept is identical to a Valve for 4.0 people, with an Action for Remy,
and close enough to an Interceptor for 3.3 people ).


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296850 13f79535-47bb-0310-9956-ffa450edef68