The quick hack I mentioned in the previous commit.
Now both 3.3 and 4.1 ( and I suppose 4.0 didn't changed since yesterday ) are
working with all 3 channels.

( httpd -X for jni channel please - shm is not yet ready :-)

That doesn't mean everything is fine - requests are forwarded but the
response is 'no context' in 4.1 ( even if the standalone is working fine ).
Still investigating.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@296856 13f79535-47bb-0310-9956-ffa450edef68