Added comments on the format used for properties.
Added method to save the properties file ( for now in a separate location ) whenever
a property is changed after startup ( by JMX or another admin application ).

Removed the need to specify the class name for the known components ( we
do know them, no need for the user to do that ). Try to load all components -
if they don't have the needed resources ( native libs, etc ) they'll just
report so and do nothing.

Made the format closer to what's used in mod_jk ( except ':' can't be used
since it's a delimiter, etc ). It seems implementing the ini format in java is
not much easier than C ...


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296864 13f79535-47bb-0310-9956-ffa450edef68