Update the config file - most of the stuff has been removed, it was just un-needed pain for the users, we knew already what was there.
The config file should be used only to override defaults, if you
have a good reason to. In addition added a ( I hope clear ) message
that the config file may be overrideni by config tools.
That will happen after I get to implement the Mbean for jk
or if another tools set properties via JkMain.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296873 13f79535-47bb-0310-9956-ffa450edef68