The java side of the changes.
For now it's just a "Hello World" sent between Apache and tomcat,
but it works fine - and that was the hard part.

Now follows the hard part, implementing a protocol that is flexible enough
to support both worker status but also config updates and other status
and config info.

Luckily there is a backup plan - to use the exact same format as in jserv
( at least to get things started )


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296883 13f79535-47bb-0310-9956-ffa450edef68