When parsing server.xml and web.xml files, use an InputSource rather than an InputStream so that we can set a system identifier.  Among other things, this allows developers to split up their input files with the use of external entities like this:
  <?xml version="1.0" encoding="ISO-8859-1"?>
  <!DOCTYPE web-app
    PUBLIC "-//Sun Microsystems, Inc.//DTD Web Application 2.3//EN"
    "http://java.sun.com/dtd/web-app_2_3.dtd"
  [
  <!ENTITY webinc PUBLIC "webinc" "webinc.xml">
  ]
  >
  <web-app>
    &webinc;
  </web-app>

Tested this with web.xml files in both unpacked directories and a packaged
WAR file, and it seems to work.

NOTE:  This patch doesn't address trying to parse tag library descriptors
that have external entities included in them.  The same approach would probably
work, but will require a couple of API changes to some private methods in
o.a.c.ContextConfig and the corresponding code in Jasper.

Based on a patch provided by Attila Szegedi <szegedia at freemail.hu> for
Bugzilla #8024.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@287967 13f79535-47bb-0310-9956-ffa450edef68