Change TYPE_CONTEXT to TYPE_PREFIX ( it's prefix map, not a context ).
Add the 'real' MAP_TYPE_CONTEXT to correspond to the webapp root, and
the TYPE_VHOST for the host.

This will allow common ( and correct ) code to deal with WEB-INF automatically,
and will let the maker set the correct vhost and context path ( that
can be used by coyote to avoid a second mapping, eventually ).


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@296975 13f79535-47bb-0310-9956-ffa450edef68