That's the big one for today. Again, the alghoritm is the one from java and from the spec, without any changes.
Virtual hosts are mapped corectly, contexts are also mapped corectly.

Still not completed - vhost aliases ( few more lines of code needed ).

Note that the current code is very sensitive to correct configuration,
I started to add some code to auto-set some of the required settings.
It is required to define each vhost and webapp - if you set a mapping
for an unexistend vhost/webapp you may get a NPE or it won't work.

( it won't be that painfull - the autoconf and better defaults
will avoid most manual settings ).

Also not completed - WEB-INF filtering and the default settings for
each webapp ( servlet/*, *.jsp ) ( there is no reason to ask the
user to configure those )


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@296980 13f79535-47bb-0310-9956-ffa450edef68