Use the WRITE_SLOT command.
Again, the idea is that the scoreboard is divided in 'slots', each slot
has a name ( identifying the 'owner' ).

We'll use the scoreboard for configuration ( at first ), with a
java bean ( used either individually or at tomcat startup ) writing
information in the scoreboard about a tomcat instance.

The lb worker will read the 'serial' number in the scoreboard header
and update it's instance table based on the content.

The synchronization is not implemented yet - if 2 tomcats
are registered at the exact same time we may get unexpected results
( a bit unlikely, will be fixed after we 'bridge' the APR locks ).


git-svn-id: https://svn.apache.org/repos/asf/tomcat@296986 13f79535-47bb-0310-9956-ffa450edef68