I believe we now extract _all_ web-server relevant information from web.xml. ( including filters, wellcome files, etc ).
Some of it is not yet supported in jk2 ( or jk1 ) - but we can try to
generate fragments for apache and other servers where this is possible
( for example Wellcome files or error pages ).

In time we can add support for those in jk2, but at least for apache
this is already implemented and it's better to use the existing impl.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@297003 13f79535-47bb-0310-9956-ffa450edef68