Add listener registration support to NamingResources, so that we'll be able to dynamically create and remove MBeans for global JNDI resources defined via the admin tool (or any other dynamic mechanism) while Tomcat is running.
Add cross reference links from each of the resource definition objects so
that they can find the NamingResources collection with which they are
associated.  At present, this is only needed to get from a
ContextResource to the corresponding ResourceParams (and vice versa), but
it's better to do things consistently across all types.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@288050 13f79535-47bb-0310-9956-ffa450edef68