Call the setter on the worker for non-specific properties.
Each channel is associated with an ajp worker ( later we can configure
that, but at this moment we use ajp13 as the main protocol ). The
load balancing parameters are set in the worker ( again, it's 1-1 relation
between worker and channel, the worker is just a protocol with an
associated channel ).


git-svn-id: https://svn.apache.org/repos/asf/tomcat@297054 13f79535-47bb-0310-9956-ffa450edef68