Update the list of properties to include more lb info.
Automatically add the worker to all the groups that are configured in.

The load balancing is now configured by including (one or many)
"group" attributes on the channel. Each group name will be used
to construct automatically an lb worker, and the channel will be
added to the lb worker.

The name of the group can then be used in uri mappings.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@297061 13f79535-47bb-0310-9956-ffa450edef68