Few more lb changes.
This is getting a bit more agressive in trying to recover
failed workers  and simplifies the logic that is used.

I tested it and seems to work very well with 'lbfactor=0' -
if the default tomcat goes down, the request goes to the
backup, when the default tomcat goes up ( after the timeout )
it'll be tried again.

As soon as shm is finished, the default will go up when it
re-register itself in the shm.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@297069 13f79535-47bb-0310-9956-ffa450edef68