Implement saving the server configuration back to "conf/server.xml", after saving the old version to "conf/server.xml.{timestamp}" as discussed on TOMCAT-DEV.
Optimize the very common cases where the <Loader> and <Manager> elements
contain all default values, and need not be written out (because the startup
process will install them automatically).

Ignore a spurious "application already installed" error when auto-deploying
context configuration files (equivalent to the existing behavior when a
WAR file or webapp directory is found in "webapps", and the corresponding
context path is already in use.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@288073 13f79535-47bb-0310-9956-ffa450edef68