Added a flag to indicate if we are loaded from jni.
Use mod_jk.so ( which includes all the functions we need, including the jni
stuff ) if in-process. ( XXX need to make this configurable or try to
detect IIS - when Nacho gets IIS working with JNI we can fix it ).

JNI code is used for in-process worker and also for unix socket and
to register tomcat.

I also removed the load() for apr, there is no way to make this work
in JDK1.4 unless we compile jkjni with -lapr -lcrypt.

Tested jkjni with all VMs I have ( except 1.1 - this is not supported with 1.1),
seesm to work fine with all.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@297081 13f79535-47bb-0310-9956-ffa450edef68