Init debug, disabled.
Use the 'debug' attribute of the uri to decide if we display information
about the mappings.

The problem is that with 'global' debug enabled we get too much info,
and that makes it hard to debug.

Using a 'log4j' aproach with a config file for all loggers is too complex.

We just have a 'debug' attribute for each component, and use it to filter.
A debug on a uri will display info about requests that are mapped to that
particular uri.

Debug on worker ( and lb worker ) will display specific info when that
worker is used.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@297102 13f79535-47bb-0310-9956-ffa450edef68