Some big changes.
I did more cleanup in the 'find_best_worker'. Right now we don't use
the 'in_recovery' or the extra delay - if the timeout expires we'll
remove the worker from 'error' state, so it can be tried again.
Normal rules are used.

If the next request is ok, great. If not - it'll go back to error state.
The only problem is that after timeout the worker will have a low lb_factor,
and will be prefered. That will be resolved by using 'int' factors and a
better mechanism for re-selection.


In addition, the scoreboard is almost completely operational - it can
do gracefull shutdown, add new workers at runtime, re-enable workers.
All this requires using some JNI and jkshm.jar - but a non-jni based
solution is in work ( for JDK1.4 and nio )


git-svn-id: https://svn.apache.org/repos/asf/tomcat@297106 13f79535-47bb-0310-9956-ffa450edef68