Started to unmangle the unix socket from apr.
If/when APR will have unix domain support - it will not need special
#ifdefs.

The channel_un is specific to unix and will be available only if HAVE_UNIX_SOCKET
is set at compile time.

( you shouldn't set it unless you want to play with the cutting-edge, there
are problems with signals sent when the other end dies - at least in JDK1.4 ).

Expect more work in this area after everything else is stable - I can't
wait to play with passing open file descriptors :-)


git-svn-id: https://svn.apache.org/repos/asf/tomcat@297107 13f79535-47bb-0310-9956-ffa450edef68