Started some work on the 1.4 based scoreboard.
This will use NIO to support the memory mapped file without requiring JNI.

If you remember, this was one of the reasons for using MsgAjp in each
shmem slot - the scoreboard will be accessed from java and C, and there's no
way to know how C structures are layed out.

We still need to 'fix' the head - since it uses a C struct at this moment.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@297112 13f79535-47bb-0310-9956-ffa450edef68