Just the empty class declaration.
It'll eventually implement the shared-memory based channel. It is
not easy, but possible - if JPDA can do it, we should be able to do it too.

Not for now, of course - it'll need a lot of work for the general case,
but we should at least better organize the Shm code.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@297113 13f79535-47bb-0310-9956-ffa450edef68