Switch the lb_factor/lb_value to int. For now we'll just use a small int ( 0..255 ) to represent the lb_value, and roll back to 0. The only role of lb_value is to make sure all workers get a similar amount of requests - with lb_factor giving more priority to some workers. There is little benefit on using floats for that, and a small int avoids the need for atomic operations.
Note that we can still enhance the model to use the number of 'active connections'
and favor the tomcat instances that have fewer active connections. We
can also use the average response time ( after we add this info :-)


Add multiple 'levels' - a genaralization of the 'local worker concept.
At init time the channels are sorted by level in different tables
( currently we have 4 levels only, hardcoded - but I doubt we need more than 2 ).
If all workers in a lower level are in error state, we move to the next.
That removes the sepcial meaning of '0',


git-svn-id: https://svn.apache.org/repos/asf/tomcat@297119 13f79535-47bb-0310-9956-ffa450edef68