Much cleaner and easier to read report, remove duplicated code and generalize a bit.
We'll display a table for each object type ( that we care about ):
- ajp workers
- uris
- endpoints

For each object we display all we can get via 'getters' ( if you want more
info, just add a new attribute for getAttribute ).

The workers are the important piece - the state and counters.

Note that at this moment we display only the info in the apache instance
that is active. It'll change to use the shmem scoreboard.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@297124 13f79535-47bb-0310-9956-ffa450edef68