Refactoring DefaultServlet to allow for common code with SSIServlet.
As far as DefaultServlet, this is a pure re-factoring with no change to the code processed.  It should be a safe change.

The new URLEncoder is very much like o.a.t.u.buf.UEncoder.  It might be good to merge the two, but by adding it for now, more people can review it.

Submitted by: Dan Sandberg x@cs.stanford.edu


git-svn-id: https://svn.apache.org/repos/asf/tomcat@288177 13f79535-47bb-0310-9956-ffa450edef68