Initial jk2 makefile.
In Apache 2 mode, or when APR is allready present
(via configure --with-apr-include/--with-apr-lib),
only jk2/jni is compiled.

In Apache 1.3 mode, when APR is available as an
external source dir (via configure --with-apr),
apr is compiled first.
In that case, we'll have to pass the proper
flags to configure.

There is still works to be done:

grab APR CFLAGS/LDFLAGS when not in Apache 2.0
mode (where apxs allready known and reports them)

handle apache 1.3 without APR

avoid using --with-apr{xxxx} in Apache 2.0....


git-svn-id: https://svn.apache.org/repos/asf/tomcat@297197 13f79535-47bb-0310-9956-ffa450edef68