Moved the reconfiguration where it belongs, in jk_config ( from worker_lb ).
This will be a generic mechansim. Added mtime to store the modified time for
the config file. We could implement reconfig by using a periodic check - but
at the moment using jkstatus is much better.

Every time jkstatus is invoked, it'll check the mtime of the config and triger
a reload if needed. To reconfigure jk2 ( well, whatever options we support -
full reconfiguration is too expensive in performance since it would require
many syncs ) you modify the config file and access jkstatus.

Other mechanism will be provided, but that should be easy enough ( and
easy to automate ).


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@297199 13f79535-47bb-0310-9956-ffa450edef68