More cosmetic changes - removing fields that are no longer used, adding comments.
Added several timing fields in the data we collect on the scoreboard,
and flags to turn on/off the collection ( per uri or globally ).

The time collection uses apr - it won't work in Apache1.3 ( or IIS ) unless
apr support is enabled.

In my test the time collection adds about 2 ms per request ( 33->35 ms average
for a simple request, with ab -c 20 ). I don't think this is much, but
if you don't use the info there's no need to collect it.

Also added a new function to workerEnv - to be called by the 'parent'
process in a multi-process server ( i.e. apache :-). This will allow
jk to acquire the scoreboard ( at least HPUX is picky about this ),
and will be used to reset the sb on startup.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@297244 13f79535-47bb-0310-9956-ffa450edef68