We can't call find_by_pid in init() because the scoreboard may not be initialized or the process may not be registered.
We can't call it in service() because linux will return the pid of the
thread, which has nothing to do with the child id ( so it won't be found
in the scoreboard ).

We need the right pid to find the childId - it used to be available in
connection_rec. We need the childId because there's no easy way to
allocate a slot in the jk scoreboard ( or detect a slot is no longer
used ).

The current workaround is to get the pid in init(), which is called ( hopefully )
in the child main thread ( which is registered in the scoreboard ).
And we find the child_num in service, when the scoreboard is supposed
to be ok.

( we do the same for apache1.3 - since child_num is only available
on the first connection )


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@297261 13f79535-47bb-0310-9956-ffa450edef68