Add support for the new DBCP Features.
Three new DBCP parameters were added:

removeAbandoned - True or false. If true abandoned db connections
are removed after the removeAbandonedTimout is exceeded if the dbcp
is nearing exhaustion.

removeAbandonedTimeout - Time in seconds since a connection was last
used before it is considered abandoned.

logAbandoned - True or false.  If true Exception stack traces are
created so that the source of an abandoned db connection can be logged.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@288213 13f79535-47bb-0310-9956-ffa450edef68