All files will compile now, with some generating empty factories if the feature is not enabled.
Use 'lb' instead of 'worker.lb' and 'ajp13' instead of 'worker.ajp13'.
( it's better to use a simpler name, and it remove another confusing use
of 'worker' name )

Remove some of the workarounds and tricks - the code is now stable and
well, it's better to fix the remaining bugs then to work around.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@297274 13f79535-47bb-0310-9956-ffa450edef68