Added a target that will build mod_jk using cpptasks from sourceforge.
This is for experimental purpose only - and to facilitate an integration.
At this moment thare are major bugs in cpptasks ( libtool doesn't work,
you have to remove history.xml, etc ), and it won't work with ant1.4
unless the .jar is in ant/lib. It won't work in ant1.5 unless my patch
is accepted.

However it's worth adding it - it's clearly good to have at least a common
dtd and semantics for compiling the c code, and cpptasks has some usefull
features.

Don't use the target unless you know what you're doing :-)


git-svn-id: https://svn.apache.org/repos/asf/tomcat@297322 13f79535-47bb-0310-9956-ffa450edef68