A bit more consistence.
We use the APACHE2_HOME env if set ( this is required in windows by the dsp ).

Load a build.properties that is specific to our OS/arch ( I have
my home nfs-mounted, and I use the same dir to build on multiple
platforms - I suspect others may be in the same situation ).

Added an arch.prefix property to be used to construct the target
dirs - again to support building multiple platfroms on the same dir.

Various fixes for hpux and solaris and win32.

Now jk2 builds fine on linux, solaris, hpux and win32.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@297336 13f79535-47bb-0310-9956-ffa450edef68