Use APR to load the DLL ( rather than adding another case for hpux ).
The major problem is Netware - the code in APR for dlopen is _very_
different from what we have in jk. If someone who uses that platform
can check if apr dlopen works fine, we can remove the commented code.

I assume the windows and standard dlopen cases are well tested, so
we'll get rid of all this platform-specific code soon.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@297342 13f79535-47bb-0310-9956-ffa450edef68