Common dispatcher for all channels - allow any jk channel to be called by the JNI code.
This is not yet completed, but will reduce the work in 1/2, since the
same code will be used on both sides and optimizations will be shared
by all components.

I hope to see in few months at least a doors channel, and maybe play
with message queues and some windows-specific things - just for the fun
of doing a small benchmark.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@297360 13f79535-47bb-0310-9956-ffa450edef68