Implement preInitCheck. As discussed on tomcat-dev.
postInitCheck is being left for a later commit (I've got to get my numbers up :).

This passes the entire testing suite, as well as any other regression test that I could think of.  So I'm fairly certain that there are no regressions.  Performance should be much the same, since it is doing the same checks as before, only in a different order.

However, now only people that care about recompiling JSPs get called, and not the other two or three that implement requestMap.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@284678 13f79535-47bb-0310-9956-ffa450edef68