Few more improvements.
- if name is not specified, construct one from the last part of the class
name, plus an int if multiple instances are present. ( can be improved ).
We should add some common 'patterns' to extract an id from a bean -
getName, getId, etc ( using introspection ).

- use findMBeanServer() instead of create - so we use the same server
as tomcat. ( assuming modeler is loaded first - I think the modeler should
do the same, first look for existing server - tomcat may run in jboss ! )

- reduce debug info

- Try to use the real class name - I don't know how it is used ( but
it shouldn't ), but it's better to see it instead of the DynamicMBeanProxy.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@297479 13f79535-47bb-0310-9956-ffa450edef68