Applied the patch supplied by Jan Luehe, to implement tag pooling/reuse based on a scheme Jan and I discussed.  Jan's note:
Tag pooling/reuse is implemented as follows:

- Every custom tag is assigned a tag handler pool, based on its full tag name
and attribute set.

- Tag handler pools are declared as servlet instance variables, so
that they can be shared across page invocations.

- Tag handler pools are instantiated with a
yet-to-be-made-configurable capacity at servlet construction time
(current default capacity is 5).

- When a tag handler is required, it is retrieved from the appropriate
tag handler pool. If the pool is empty, a new tag handler is instantiated and
returned.

- When a tag handler is no longer needed, it is returned to the appropriate
pool. It is added to the pool, unless the pool has already reached its
capacity, in which case the tag handler's release() method is called.

- The servlet's destroy() method enumerates the servlet's tag handler
pools and calls the release() method of every available tag handler.


Outstanding issues:
- Make tag handler pool capacity configurable.
- Optimize tag reuse in the body of iteration tags.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@304933 13f79535-47bb-0310-9956-ffa450edef68