Teach URL how to parse userInfo.
This allows URL to parse a string of the form:
ftp://user:pass@ftp.mycompany.com/

This actually doesn't fix any Tomcat problems (assuming that the string is correct to start with), and this class is only used by TC33. I just sleep easier if the URL is parsed correctly.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@297539 13f79535-47bb-0310-9956-ffa450edef68