Time to break Jasper2...
This is the first round of refactoring, eliminating the huge duplication
between JspEngineContext and CLContext.

JspCompilationContext is now a normal bean, with getters and setters. It
can be overriden or used as is, and it includes all the code that used to
be in JspEngineContext for mangling.

The main purpose of the class is to deal with paths and resources -
including mangling, etc.

CommandLineContext is no longer needed - we use the same code as in JspEngineContext.
I moved some of the extra checks to deal with the lack of RuntimeContext
when doing CLI compilation.

There are few methods remaining in JspEngineContext - dealing with class
loading. The goal is to move them back to the base class, and thus
enable JspC to do the same depenency checking as JspServlet.

( at the moment JspC can't look for include deps )


git-svn-id: https://svn.apache.org/repos/asf/tomcat@304956 13f79535-47bb-0310-9956-ffa450edef68