Second stage of braking jasper2.
I split some of the huge methods in smaller ( and hopefully cleaner )
ones, isolated the CLI processing at the end of the file.

The class can be used as an ant task ( as in the javadoc comments ).

I also added calls to isOutDated(), so <jasper2> will check if the
.java class is newer. Looking for includes is missing for now.

Some features are lost: it is no longer possible to compile to webapps
at once, and the mangler is identical with the one used by JspServlet.

Everything else should work as before ( emphasis on 'should' ). I run
few tests and it seems fine.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@304957 13f79535-47bb-0310-9956-ffa450edef68