2 bugs: - wrong POST body ( due to changes to support JNI, we reuse the request buffer for the response - there is no need to use 2 buffers in the roundtrip ).
- if a single worker is used ( typical case ), on some cases send() doesn't detect
that tomcat was restarted ( i.e. doesn't return an error even if the other end is
down ). We get the error only when we do the first read.

The change also deals better with recovery when we have a POST ( previously
this didn't recover ). The 'recoverable' field is used to indicate if any
response processing happened - if all we did was send the request head and eventually
the initial body chunk, but the first read from tomcat fails, we'll close the
connection and try again.

It is still not perfect, I'm still working on that.

No idea why write() ( or send() ) are successfull. Well, probably TCP related, I'll
 take a look with etherreal.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@297557 13f79535-47bb-0310-9956-ffa450edef68