Initial version for the proposed 2-layer config tool.
The main focus is to deal with storing the configuration and
supporting the current bean-model of tomcat as well as JMX components.

The Config interface is all that a user ( i.e. management app or
Main/Catalina/Startup component that sets up tomcat ) will use.

Internally it'll use jdk1.4 prefs / JNDI / XML / INI / Properties
to read/store the data, and the current Bean introspection or JMX
to configure the runtime.

Any change to the runtime must be done via Config or JMX ( Config
will use the JMX notifications to intercept the settings on JMX ).
Config will persist the user changes ( and only user changes, as
oposed to calling all getters ).

In addition I would like to support the ${} properties ( as supported
by tomcat3.3 ), and have the storage use ${} expressions.

Please - review and help !


git-svn-id: https://svn.apache.org/repos/asf/tomcat@297663 13f79535-47bb-0310-9956-ffa450edef68