If a Servlet generates a RuntimeException, it is wrapped by a ServletException.  Since there is an entry for ServletException in the web.xml, the actual root cause (in this case, an IllegalStateException) was never unwrapped, thus leading to unexpected behavior.
I made a simple modification to the logic so that if the throwable passed to the ErrorDispatcherValve is a ServletException, it will attempt to get the root cause.  If the root cause is null, then proceed with the error page lookup with the ServletException as the throwable object, otherwise, use the root cause for the error page lookup.
Submitted by:	Ryan Lubke (ryan.lubke@sun.com)


git-svn-id: https://svn.apache.org/repos/asf/tomcat/container/catalina/trunk@301155 13f79535-47bb-0310-9956-ffa450edef68