Switch to commons-logging instead of jasper's own logging.
To enable debugging - use log4j.properties or the jdk1.4 config.
Right now it uses the 'class name' convention - i.e. the name of
the logger where you can enable debug is the class name.

We could use a single logger for all jasper.

I also added code to log the translation and compilation times -
if they are more than 500ms. It can be removed before release ( or
made configurable ), but for now it's good to know.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@305113 13f79535-47bb-0310-9956-ffa450edef68