Switch to commons-logging.
While looking at the code, it seems one thread per webapp is used to
find updates - that doesn't scale very well. A single thread should
be used for all webapps.

Also looking at the code, the MANIFEST stuff is interesting but it
would be nice to make it more generic. I'm planning to try to merge
some of this code into discovery, and allow more flexibility in
finding jars and resolving dependencies. The code would be perfect
for ant and other projects.

I added a bit of code to support the case when the parent loader
is not a URLClassLoader - i.e. if we start from ant. Any classloader
that has a getClassPath method can be used. Now compiling jsps when
started from ant should work/


git-svn-id: https://svn.apache.org/repos/asf/tomcat@301178 13f79535-47bb-0310-9956-ffa450edef68