- Patch by Mark Roth
Attached is a patch to the jakarta-servletapi-5 workspace with various
examples that exercise new features in JSP 2.0.  I've also fixed some of
the JSP 1.2 examples (the paths were not correctly brought over when the
examples were moved to the jakarta-servletapi-5 workspace).  I've also
updated some of the XML Schemas in the jsr154/ subdirectory which had
stale versions of the jsr152 schemas.

PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/tomcat@266561 13f79535-47bb-0310-9956-ffa450edef68