More 'fast-build' targets.
I intend to remove the compilation part of build2.xml, and move
the rest in other places ( ant-launcher.xml ? ). While it's very
fast, it is also too hard to maintain.

The current solution is a bit slower, but it uses the original
build.xml files for each component.

What I'm doing is simply call the build.xml with the build
dir and .jar target pointing to our build and the final destination.
That avoids file copy.
In addition it uses <uptodate> to avoid recompiling what doesn't
change.

After a full build, I get a second one in ~10-15 sec, and then it
seems to work fine and compile only what changed.
It is also possible to call individual targets.

That's for people with <2GHz processors only :-)


git-svn-id: https://svn.apache.org/repos/asf/tomcat@304225 13f79535-47bb-0310-9956-ffa450edef68