Merged the startup code from build2.xml. The 3 startup methods are independent ( launcher, java exec and task ).
Of particular interest ( I think ) are the 2 run targets.
"run" will load catalina in-process and _return_ imediately after
the init block is done - i.e. you no longer need the sleep.

( same thing that is used for example by anteater to start tomcat33).

This is still not completed - and will need to be integrated with the
launcher.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/container/catalina/trunk@301302 13f79535-47bb-0310-9956-ffa450edef68