Use the catalina.properties file to customize the package protection/access. This new security m echanism enable the customization, at runtime, of which package should be protected.
the following package will be protected by default:

o.a.catalina
o.a.jasper(*)
o.a.coyote
o.a.tomcat.util

(*) Tomcat 5 is broken when a JSP use a class from jsp20el.jar and when the SecurityManager is t
urned on. Even if you remove all the protection, Tomcat fail to properly runs the example.


o.a.coyote.tomcat5 has been securized in order to support package protection.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/container/catalina/trunk@301311 13f79535-47bb-0310-9956-ffa450edef68