Use commons-logging.
Add a short message for servlets that take a long time to load
( right now it's 200 ms - we can increase it, but no need to
make it configurable, it's mostly do debug perf. ).

I used idea to edit the file - and it removed unused imports and
forced me to fix some name mismatches... Probably should go back
to emacs...


git-svn-id: https://svn.apache.org/repos/asf/tomcat@301318 13f79535-47bb-0310-9956-ffa450edef68