Second attempt.
Restore the flushBuffer and clean the error.

I added a quick test for new exception - if they are different
it most likely means an error in the error page - and servlet
error handiling should take care of it.

Whatever we set is cleaned up. Probably we should restore with what
was before - but I'm not sure.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@305317 13f79535-47bb-0310-9956-ffa450edef68