Bug fix for BUG #13364
A Web Application Context reload by the manager web application
was causing named JNDI resources to disappear.

A webapp reload needs to dump the webapp classloader, then
recreate. The CL is bound to the naming context so the
reload was issing a NamingContext STOP_EVENT and then a
START_EVENT.  This removed all the JNDI named resources
but the code which runs at webapp startup which creates
the JNDI named resources is not run on a reload.

I fixed this by removing the START and STOP events and
adding BEFORE_STOP_EVENT and AFTER_START_EVENT
lifecycle events whose only purpose is to bind or unbind the
ClassLoader to the JNDI context.

Defined JNDI resources are now preserved on a web app reload.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@288612 13f79535-47bb-0310-9956-ffa450edef68