Same fix for jk2. Don't send initial chunk for chunked encoding.
The reason the data is sent for POST - in almost all cases tomcat
will read it ( required by the spec - to process the data ). This way
we avoid a round-trip.

We don't send more than 8k for few reasons - tomcat may want to
discard the request ( for example to prevent a huge body that it
can't handle ), the TCP stack would do round-trips anyway ( for flow-control )
and most requests are <8k anyway.

There is a major problem ( since the first release of Ajp13) related
with the fact that this chunk of data is not properly
marshalled ( it doesn't have the 'type' descriptor ).


git-svn-id: https://svn.apache.org/repos/asf/tomcat@298163 13f79535-47bb-0310-9956-ffa450edef68