This patch allows for the following: %{xxx}i  xxx is the name of the incoming header %{xxx}c  xxx is the name of a specific cookie %{xxx}r  xxx is an attribute in the ServletRequest %{xxx}s  xxx is an attribute in the HttpSession
For %A - uses InetAddress.getLocalHost() instead of hardcoded "127.0.0.1"

Patch also allows access log not be rotated. This is done by adding a new
attribute called rotatable. It defaults to true. If set to false - then the
access log does not rotate. (Which will make a lot of tomcat-users happy)

Submitted by Tim Funk <funkman@joedog.org>


git-svn-id: https://svn.apache.org/repos/asf/tomcat@288625 13f79535-47bb-0310-9956-ffa450edef68